Flesh out AWS CLI specfile (#14)
It now uses the pyproject macros which is the
new standard and is needed to build the CLI.

It also required a patch to work with the dependenices
currently in Fedora:

* Update the setup.cfg to loosen the version range of the
  packages required by the CLI
* Update usage of promptoolkit usage since its pin required
  bumping up to a new major version

We should be able to get rid of this patch once the version
ranges are updated upstream. Most of these should be able to be
updated.