Move aws-c-* executables to standard package
Fixes #34. This was accomplished by making the standard package
rely on the libs package and moving the executable to the
standard package. This was chosen over just moving all of the files
from the libs package to the standard package because it makes it
more consistent naming when declaring cross-CRT dependencies.
Furthermore none of the CRT modules need these executables to
function properly.